feat(exactMatch): add exactMatchSwitch Added a new toggle for enabling exact-match filtering from the TagDrawer.  This returns only images that have exactly the tags specified in the TagDrawer, including no tag selection returning images that have no tags applied at all.
Commit comes as is, author offers no warranty and assumes no liability in the event of catastrophic failure.  (Okay, no, it does work... I just had a lot of brain fog when working on it and so I have low confidence.  You know what will help with that?  TESTS!  Coming soon.)